Move bash directory to shell, for hybrid setup
Summary:

I want to start moving toward a setup that will tolerate me working
in multiple shells (in particular both bash and zsh). To that end,
I want to call the section of my dotconfar `shell` rather than `bash`.

The first thing I want to add to it is my starship.rs config.